+++ core/common/gnuid.cpp
@@ -37,8 +37,8 @@ void GnuID::encode(Host *h, const char *salt1, const char *salt2, unsigned char 
     }
 
     int s1 = 0, s2 = 0;
-    int n = ((std::max)(salt1?strlen(salt1):0, salt2?strlen(salt2):0)/16 + 1) * 16;
-    for (int i = 0; i < n; i++)
+    size_t n = ((std::max)(salt1?strlen(salt1):0, salt2?strlen(salt2):0)/16 + 1) * 16;
+    for (size_t i = 0; i < n; i++)
     {
         unsigned char ipb = id[i%16];
