+++ core/common/http.cpp
@@ -233,7 +233,7 @@ void HTTP::send(const HTTPResponse& response)
     writeLine("");
 
     if (response.body.size())
-        write(response.body.data(), response.body.size());
+        write(response.body.data(), static_cast<int>(response.body.size()));
 }
 
 // -----------------------------------
