+++ core/common/mkv.cpp
@@ -24,7 +24,7 @@ void MKVStream::sendPacket(ChanPacket::TYPE type, const byte_string& data, bool 
     ChanPacket pack;
     pack.type = type;
     pack.pos  = ch->streamPos;
-    pack.len  = data.size();
+    pack.len  = static_cast<unsigned>(data.size());
     pack.cont = continuation;
     memcpy(pack.data, data.data(), data.size());
 
@@ -39,7 +39,7 @@ void MKVStream::sendPacket(ChanPacket::TYPE type, const byte_string& data, bool 
 
 bool MKVStream::hasKeyFrame(const byte_string& cluster)
 {
-    MemoryStream in(const_cast<unsigned char*>(cluster.data()), cluster.size());
+    MemoryStream in(const_cast<unsigned char*>(cluster.data()), static_cast<int>(cluster.size()));
 
     VInt id   = VInt::read(in);
     VInt size = VInt::read(in);
@@ -55,7 +55,7 @@ bool MKVStream::hasKeyFrame(const byte_string& cluster)
 
         if (id.toName() == "SimpleBlock")
         {
-            MemoryStream mem((void*)blockData.data(), blockData.size());
+            MemoryStream mem((void*)blockData.data(), static_cast<int>(blockData.size()));
 
             VInt trackno = VInt::read(mem);
             if (trackno.uint() == m_videoTrackNumber)
@@ -95,7 +95,7 @@ void MKVStream::rateLimit(uint64_t timecode)
     // Timecode は単調増加ではないが、少しのジッターはバッファーが吸収
     // してくれるだろう。
 
-    unsigned int secondsFromStart = timecode * m_timecodeScale / 1000000000;
+    unsigned int secondsFromStart = static_cast<unsigned>(timecode * m_timecodeScale / 1000000000);
     unsigned int ctime = sys->getTime();
 
     if (m_startTime + secondsFromStart > ctime) // if this is into the future
@@ -122,7 +122,7 @@ void MKVStream::sendCluster(const byte_string& cluster, Channel* ch)
             continuation = false;
     }
 
-    MemoryStream in((void*) cluster.data(), cluster.size());
+    MemoryStream in((void*) cluster.data(), static_cast<int>(cluster.size()));
 
     VInt id   = VInt::read(in);
     VInt size = VInt::read(in);
@@ -165,7 +165,7 @@ void MKVStream::sendCluster(const byte_string& cluster, Channel* ch)
             size_t pos = 0;
             while (pos < buffer.size())
             {
-                int next = (std::min)(pos + 15*1024, buffer.size());
+                int next = static_cast<int>((std::min)(pos + 15*1024, buffer.size()));
                 sendPacket(ChanPacket::T_DATA, buffer.substr(pos, next-pos), continuation, ch);
                 continuation = true;
                 pos = next;
@@ -174,7 +174,7 @@ void MKVStream::sendCluster(const byte_string& cluster, Channel* ch)
         } else {
             buffer += id.bytes + size.bytes;
             buffer.append(payload.begin(), payload.end());
-            MemoryStream mem((void*)buffer.c_str(), buffer.size());
+            MemoryStream mem((void*)buffer.c_str(), static_cast<int>(buffer.size()));
             mem.rewind();
             VInt id = VInt::read(mem);
         }
@@ -201,7 +201,7 @@ void MKVStream::readTracks(const std::string& data)
 
         if (id.toName() == "TrackEntry")
         {
-            int end = mem.getPosition() + size.uint();
+            int end = mem.getPosition() + static_cast<int>(size.uint());
             int trackno = -1;
             int tracktype = -1;
 
@@ -215,7 +215,7 @@ void MKVStream::readTracks(const std::string& data)
                 else if (id.toName() == "TrackType")
                     tracktype = (uint8_t) mem.readChar();
                 else
-                    mem.skip(size.uint());
+                    mem.skip(static_cast<int>(size.uint()));
             }
 
             if (tracktype == 1)
@@ -225,7 +225,7 @@ void MKVStream::readTracks(const std::string& data)
             }
         }else
         {
-            mem.skip(size.uint());
+            mem.skip(static_cast<int>(size.uint()));
         }
     }
 }
