+++ core/common/stream.h
@@ -208,11 +208,11 @@ public:
     void    writeString(const char *);
     void    writeString(const std::string& s)
     {
-        write(s.data(), s.size());
+        write(s.data(), static_cast<int>(s.size()));
     }
     void    writeString(const String& s)
     {
-        write(s.c_str(), s.size());
+        write(s.c_str(), static_cast<int>(s.size()));
     }
     void    writeStringF(const char *, ...) __attribute__ ((format (printf, 2, 3)));
 
@@ -266,8 +266,8 @@ public:
             double tdiff = now - m_lastUpdate;
             if (tdiff >= 1.0)
             {
-                m_bytesInPerSec     = (m_totalBytesIn - m_lastBytesIn) / tdiff;
-                m_bytesOutPerSec    = (m_totalBytesOut - m_lastBytesOut) / tdiff;
+                m_bytesInPerSec     = static_cast<unsigned>((m_totalBytesIn - m_lastBytesIn) / tdiff);
+                m_bytesOutPerSec    = static_cast<unsigned>((m_totalBytesOut - m_lastBytesOut) / tdiff);
 
                 m_bytesInPerSecAvg  = Exp * m_bytesInPerSecAvg + (1-Exp) * m_bytesInPerSec;
                 m_bytesOutPerSecAvg = Exp * m_bytesOutPerSecAvg + (1-Exp) * m_bytesOutPerSec;
@@ -288,8 +288,8 @@ public:
         unsigned int    bytesInPerSec() { update(); return m_bytesInPerSec; }
         unsigned int    bytesOutPerSec() { update(); return m_bytesOutPerSec; }
 
-        unsigned int    bytesInPerSecAvg() { update(); return m_bytesInPerSecAvg; }
-        unsigned int    bytesOutPerSecAvg() { update(); return m_bytesOutPerSecAvg; }
+        unsigned int    bytesInPerSecAvg() { update(); return static_cast<unsigned>(m_bytesInPerSecAvg); }
+        unsigned int    bytesOutPerSecAvg() { update(); return static_cast<unsigned>(m_bytesOutPerSecAvg); }
 
         unsigned int    m_totalBytesIn, m_totalBytesOut;
         unsigned int    m_lastBytesIn, m_lastBytesOut;
@@ -494,7 +494,7 @@ public:
     {
         if (buf.size() > 0)
         {
-            stream->write(buf.c_str(), buf.size());
+            stream->write(buf.c_str(), static_cast<int>(buf.size()));
             buf.clear();
         }
     }
