+++ core/common/xml.cpp
@@ -256,7 +256,7 @@ int XML::Node::findAttrID(const char *name)
 // ----------------------------------
 char *XML::Node::findAttr(const char *name)
 {
-    int nlen = strlen(name);
+    size_t nlen = strlen(name);
     for (int i=1; i<numAttr; i++)
     {
         char *an = getAttrName(i);
@@ -283,17 +283,17 @@ void XML::Node::write(Stream &out, int level)
     char *name = getAttrValue(0);
 
     out.write("<", 1);
-    out.write(name, strlen(name));
+    out.write(name, static_cast<int>(strlen(name)));
 
     for (int i=1; i<numAttr; i++)
     {
         out.write(" ", 1);
         char *at = getAttrName(i);
-        out.write(at, strlen(at));
+        out.write(at, static_cast<int>(strlen(at)));
 
         out.write("=\"", 2);
         char *av = getAttrValue(i);
-        out.write(av, strlen(av));
+        out.write(av, static_cast<int>(strlen(av)));
         out.write("\"", 1);
     }
 
@@ -305,7 +305,7 @@ void XML::Node::write(Stream &out, int level)
         out.write(">\n", 2);
 
         if (contData)
-            out.write(contData, strlen(contData));
+            out.write(contData, static_cast<int>(strlen(contData)));
 
         if (child)
             child->write(out, level+1);
@@ -314,7 +314,7 @@ void XML::Node::write(Stream &out, int level)
             out.write(tabs, strlen(tabs));
 #endif
         out.write("</", 2);
-        out.write(name, strlen(name));
+        out.write(name, static_cast<int>(strlen(name)));
         out.write(">\n", 2);
     }
